+++ src/cf-expandables/src/ExpandableTransition.js
@@ -1,7 +1,6 @@
 // Required modules.
 const Events = require( 'cf-atomic-component/src/mixins/Events.js' );
 const BaseTransition = require( 'cf-atomic-component/src/utilities/transition/BaseTransition' );
-const contains = require( 'cf-atomic-component/src/utilities/dom-class-list' ).contains;
 
 // Exported constants.
 const CLASSES = {
@@ -38,7 +37,7 @@ function ExpandableTransition( element ) {
       _transitionComplete.bind( this )
     );
 
-    if ( contains( element, CLASSES.OPEN_DEFAULT ) ) {
+    if ( element.classList.contains( CLASSES.OPEN_DEFAULT ) ) {
       this.expand();
     } else {
       this.collapse();
@@ -52,13 +51,13 @@ function ExpandableTransition( element ) {
    * Handle the end of a transition.
    */
   function _transitionComplete() {
-    if ( contains( element, CLASSES.EXPANDED ) ) {
+    if ( element.classList.contains( CLASSES.EXPANDED ) ) {
       this.dispatchEvent( 'expandEnd', { target: this } );
 
       if ( element.scrollHeight > previousHeight ) {
         element.style.maxHeight = element.scrollHeight + 'px';
       }
-    } else if ( contains( element, CLASSES.COLLAPSED ) ) {
+    } else if ( element.classList.contains( CLASSES.COLLAPSED ) ) {
       this.dispatchEvent( 'collapseEnd', { target: this } );
     }
   }
@@ -68,7 +67,7 @@ function ExpandableTransition( element ) {
    * @returns {ExpandableTransition} An instance.
    */
   function toggleExpandable() {
-    if ( contains( element, CLASSES.COLLAPSED ) ) {
+    if ( element.classList.contains( CLASSES.COLLAPSED ) ) {
       this.expand();
     } else {
       this.collapse();