+++ CHANGELOG.md
@@ -18,6 +18,8 @@ We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 
 ### Removed
 - **cf-atomic-component:** [PATCH] Removes deprecated function bind polyfill.
+- **cf-expandables:** [MINOR] Remove classlist-polyfill for IE9.
+- **cf-atomic-component:** [MINOR] Remove dom-class-list module for IE9.
 
 ## 7.3.0 - 2018-11-06